+++ lib/selenium_record.rb
@@ -1,4 +1,5 @@
-Dir["#{File.dirname(__FILE__)}/selenium_record/*.rb"].each { |f| require f }
+require_relative 'selenium_record/base'
+require_relative 'selenium_record/navigation_item'
 
 # Selenium Record provides objects for interacting with browser and managing
 # the application during tests. Also joins in every object behaviour and